Move auth API types to a separate file
